+++ torch/csrc/jit/codegen/cuda/test/test_gpu_utils.cpp
@@ -42,7 +42,7 @@ TEST_F(NVFuserTest, FusionMergeDims_CUDA) {
       {p[0], p[1], p[2], p[3], p[4], p[5], p[6], p[7], p[8], p[9], p[10]});
   std::vector<size_t> dims{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
   auto merged = scheduler_utils::mergeDims(tv, {2, 3, 7, 8, 9}, dims);
-  TORCH_CHECK(merged == 2);
+  TORCH_CHECK(merged == (size_t)2);
   std::vector<int64_t> expect_shape{
       p[0], p[1], p[2] * p[3] * p[7] * p[8] * p[9], p[4], p[5], p[6], p[10]};
   TORCH_CHECK(tv->nDims() == expect_shape.size());
